Changes made:1. Updated Banner 2. Compose File updated with the fix to #68
Root cause of the service not to start was container wasn't being able to interact with the mysql container running in local environment.
Due to the limitation of docker linux inter container communication & port blocking issue. Though the port was opened for the connection to outerworld, yet couldn't be done, since there wasn't the port resolution/port forwarding for incomming connection. To resolve this docker-host image was added in compose file. It takes the incoming requests & facilitate the process by opening a communication bridge between containers.\n

closes #68